# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job Application Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resume-versions:
    get:
      operationId: listResumeVersions
      tags: [resume-versions]
      summary: List resume versions
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of resume versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResumeVersion"
  /resume-versions/{id}:
    get:
      operationId: getResumeVersion
      tags: [resume-versions]
      summary: Get a resume version by ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Resume version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeVersion"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateResumeVersion
      tags: [resume-versions]
      summary: Update a resume version
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResumeVersionBody"
      responses:
        "200":
          description: Updated resume version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeVersion"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteResumeVersion
      tags: [resume-versions]
      summary: Delete a resume version
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /resume-versions/{id}/approve:
    post:
      operationId: approveResumeVersion
      tags: [resume-versions]
      summary: Approve a resume version
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalEvaluationBody"
      responses:
        "200":
          description: Approved resume version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeVersion"
        "404":
          $ref: "#/components/responses/NotFound"
  /resume-versions/{id}/reject:
    post:
      operationId: rejectResumeVersion
      tags: [resume-versions]
      summary: Reject a resume version
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalEvaluationBody"
      responses:
        "200":
          description: Rejected resume version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeVersion"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Cover Letter Versions ─────────────────────────────────────────────────
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Role Profile ────────────────────────────────────────────────────────
...
    ResumeVersion:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        baseResumeVersionId:
          type: ["integer", "null"]
        label:
          type: ["string", "null"]
        templateId:
          type: ["string", "null"]
        status:
          type: string
        tailoredDocumentText:
          type: ["string", "null"]
        tailoredBullets:
          description: Array of bullet objects with claim attribution
        diffData:
          nullable: true
          description: Structured diff showing added/removed/reordered bullets
        claimIds:
          type: array
          items:
            type: integer
        fileUrl:
          type: ["string", "null"]
        rawContent:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - claimIds
        - createdAt
        - updatedAt
    UpdateResumeVersionBody:
      type: object
      properties:
        label:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        rawContent:
          type: ["string", "null"]

    # ── Cover Letter Version ─────────────────────────────────────────────────
...
    ApprovalEvaluationBody:
      type: object
      properties:
        rubric:
          type: ["object", "null"]
          properties:
            truthfulnessScore:
              type: ["integer", "null"]
            relevanceScore:
              type: ["integer", "null"]
            formattingScore:
              type: ["integer", "null"]
            attributionScore:
              type: ["integer", "null"]
        editDistance:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...