# lib/api-zod/src/generated/types/resumeVersion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Application Operations Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeVersion {
  id: number;
  /** @nullable */
  jobId?: number | null;
  /** @nullable */
  baseResumeVersionId?: number | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  templateId?: string | null;
  status: string;
  /** @nullable */
  tailoredDocumentText?: string | null;
  /** Array of bullet objects with claim attribution */
  tailoredBullets?: unknown;
  /** Structured diff showing added/removed/reordered bullets */
  diffData?: unknown | null;
  claimIds: number[];
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  rawContent?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}