# lib/db/src/schema/resume-versions.ts
import {
  pgTable,
  text,
  serial,
  timestamp,
  integer,
  jsonb,
  index,
} from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { jobsTable } from "./jobs";
import { baseResumeVersionsTable } from "./base-resume-versions";

/**
 * Resume versions — AI-tailored resume drafts awaiting human approval.
 *
 * Each version is generated by the resume tailor pipeline and starts in
 * `pending_approval` state. The user reviews the diff (added/removed/reordered
 * bullets) in the dashboard, decides on each change individually, then approves
 * or rejects the version. Per-change decisions are serialised into the `notes`
 * field before the approval call.
 *
 * Approved versions can be linked to applications via `applications.resumeVersionId`.
 *
 * State machine: `pending_approval` → `approved` | `rejected`
 * Any attempt to approve/reject from a non-`pending_approval` state returns HTTP 409.
 */
export const resumeVersionsTable = pgTable(
  "resume_versions",
  {
    /** Auto-incrementing primary key. */
    id: serial("id").primaryKey(),

    /** The job this resume was tailored for. Cascade-deletes the version if the job is deleted. */
    jobId: integer("job_id").references(() => jobsTable.id, {
      onDelete: "cascade",
    }),

    /**
     * The exact base resume version used as the source-of-truth input when this
     * tailored draft was generated.
     */
    baseResumeVersionId: integer("base_resume_version_id").references(
      () => baseResumeVersionsTable.id,
      {
        onDelete: "set null",
      },
    ),

    /**
     * Canonical AI lineage key for this generated artifact.
     * Nullable during rollout so legacy drafts remain diagnosable but are not
     * considered valid lineage.
     */
    runId: text("run_id"),

    /**
     * Event-log row for the originating AI run or downstream lineage anchor.
     * Nullable-first for legacy rows created before the canonical contract.
     */
    eventLogId: integer("event_log_id"),

    /** Human-readable label (e.g. "AI tailored — Jun 15, 2025"). Auto-set by the pipeline. */
    label: text("label"),

    /** Built-in or admin-provided renderer template used for this resume draft. */
    templateId: text("template_id"),

    /**
     * Current state. `pending_approval` until the user acts.
     * Values: `pending_approval`, `approved`, `rejected`
     */
    status: text("status").notNull().default("pending_approval"),

    /** Full tailored plain-text resume draft generated from the base resume. */
    tailoredDocumentText: text("tailored_document_text"),

    /**
     * Array of validated tailored bullet objects.
     * Each entry shape: `{ text: string, claimIds: number[], section: string, isAggregated: boolean, originalText: string | null }`
     * All claimIds are validated against the Claims Ledger before storage.
     */
    tailoredBullets: jsonb("tailored_bullets").notNull().default([]),

    /**
     * Structured diff produced by the pipeline.
     * Shape: `{ addedBullets: string[], removedBullets: string[], reorderedSections: string[], summary: string,
     *           generatedAt: string, modelName: string, bulletsTotal: number, bulletsPassedValidation: number, bulletsDiscarded: number }`
     */
    diffData: jsonb("diff_data"),

    /** Flat list of all claim IDs referenced by this version (union of all bullet claimIds). */
    claimIds: integer("claim_ids").array().notNull().default([]),

    /** URL to the exported PDF/DOCX file (if the version has been rendered). */
    fileUrl: text("file_url"),

    /** Raw AI output stored when JSON parsing fails (for debugging and regeneration). */
    rawContent: text("raw_content"),

    /**
     * User or system notes. When the dashboard approval flow is used, per-change
     * review decisions (accepted/rejected bullets) are serialised here before the
     * approve call fires, providing a traceable record of the review.
     */
    notes: text("notes"),

    createdAt: timestamp("created_at", { withTimezone: true })
      .notNull()
      .defaultNow(),
    updatedAt: timestamp("updated_at", { withTimezone: true })
      .notNull()
      .defaultNow()
      .$onUpdate(() => new Date()),
  },
  (table) => [
    index("resume_versions_job_id_idx").on(table.jobId),
    index("resume_versions_base_resume_version_id_idx").on(
      table.baseResumeVersionId,
    ),
    index("resume_versions_run_id_idx").on(table.runId),
    index("resume_versions_event_log_id_idx").on(table.eventLogId),
    index("resume_versions_status_idx").on(table.status),
  ],
);

/** Zod schema for inserting a resume version (omits server-managed fields). */
export const insertResumeVersionSchema = createInsertSchema(
  resumeVersionsTable,
).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
});

/** Type for a new resume version insert payload. */
export type InsertResumeVersion = z.infer<typeof insertResumeVersionSchema>;

/** Type for a full resume version row as returned from the database. */
export type ResumeVersion = typeof resumeVersionsTable.$inferSelect;